refactor(schema): Rename secret's schemas and types file
